Remove all unneeded mash functions
